Display service form in two columns
The first column contains the 'general' form fields like the name,
team, crm id and some config flags. The second column is filled
with the service status radio groups.

This improves the form visually and makes it easier to perform
tasks on.

In addition to this change, the global setup of the form was updated
sligthly. The functionality that was reused by the edit and create
forms has been abstracted into a form type of its own. This typ is
then extended by the CreateServiceType and EditServiceType.
