+++ src/Surfnet/ServiceProviderDashboard/Infrastructure/DashboardBundle/Form/Service/CreateServiceType.php
@@ -1,0 +1,37 @@
+<?php
+
+/**
+ * Copyright 2017 SURFnet B.V.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+namespace Surfnet\ServiceProviderDashboard\Infrastructure\DashboardBundle\Form\Service;
+
+use Surfnet\ServiceProviderDashboard\Application\Command\Service\CreateServiceCommand;
+use Symfony\Component\OptionsResolver\OptionsResolver;
+
+class CreateServiceType extends ServiceType
+{
+    public function configureOptions(OptionsResolver $resolver)
+    {
+        $resolver->setDefaults(array(
+            'data_class' => CreateServiceCommand::class,
+        ));
+    }
+
+    public function getBlockPrefix()
+    {
+        return 'dashboard_bundle_service_type';
+    }
+}
