+++ src/Surfnet/ServiceProviderDashboard/Infrastructure/DashboardBundle/Resources/views/Service/create.html.twig
@@ -1,18 +1,5 @@
-{% extends '::base.html.twig' %}
-
 {% form_theme form 'form/fields.html.twig' %}
 
-{% block page_heading %}{{ 'service.create.title'|trans }}{%endblock%}
+{% extends 'DashboardBundle:Service:base_service.html.twig' %}
 
-{% block body %}
-
-    {% for message in app.flashes('error') %}
-        <div class="message error">
-            {{ message }}
-        </div>
-    {% endfor %}
-
-
-    {{ form(form, {'attr': {'novalidate': 'novalidate'}}) }}
-
-{% endblock %}
+{% block page_heading %}{{ 'service.create.title'|trans }}{% endblock %}
